% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliarFunctions.R
\name{AllContrasts}
\alias{AllContrasts}
\title{Constructs a matrix with all the contrasts for pairwise comparisons.}
\usage{
AllContrasts(n)
}
\arguments{
\item{n}{an integer number larger than 1 giving the number of contrasts
defining the pairwise comparisons pairwise comparisons.}
}
\value{
a matrix of dimension n(n-1)/2 x n.
}
\description{
Constructs a matrix with all the contrasts for pairwise comparisons.
}
\details{
This is an auxiliar function forming a contrast matrix of all
possible. Generates an error if n is smaller than 2.
}
\examples{
AllContrasts(3)
}
\author{
Rodrigo Labouriau
}
