% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliarFunctions.R
\name{ApproxWaldPvalues}
\alias{ApproxWaldPvalues}
\title{Wald approximation for the p-values of pairwise comparisons based on the design matrix}
\usage{
ApproxWaldPvalues(Effects, CovMatrix, DesignMatrix, padjust = "fdr")
}
\arguments{
\item{Effects}{a vector containing the effects}

\item{CovMatrix}{the covariance matrix of the effects}

\item{DesignMatrix}{design matrix}

\item{padjust}{method for correcting for multiple testing as in the
function p.adjust (default = "fdr", if NULL no adjustments is made)}
}
\value{
a vector of p-values.
}
\description{
Wald approximation for the p-values of pairwise comparisons based on the design matrix
}
\author{
Rodrigo Labouriau
}
