% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods4PostHoc.R
\name{plot.PostHoc}
\alias{plot.PostHoc}
\title{Plot method for objects of class PostHoc}
\usage{
\method{plot}{PostHoc}(x, y, ...)
}
\arguments{
\item{x}{an object of class PostHoc to be printed.}

\item{y}{an object of class PostHoc to be printed.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
none
}
\description{
Plot method for objects of class PostHoc
}
\examples{
MM <- glm(Y ~ Treatment + 0,  data = DeIdentifiedExample)
GG <- posthoc(MM)
plot(GG)
}
\author{
Rodrigo Labouriau
}
