% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods4PostHoc.R
\name{print.PostHoc}
\alias{print.PostHoc}
\title{Print methods for objects of class PostHoc}
\usage{
\method{print}{PostHoc}(x, digits = 4, ...)
}
\arguments{
\item{x}{an object of class PostHoc to be printed}

\item{digits}{number of digits in the output (default = 4)}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a dataframe with two variables, Levels a factor containing the levels
 of the effects and ParaeterCI which is a factor with the effects and the
 corresponding confidence intervals and the grouping combined in a character
 constructed in such a way that when printing this dataframe yields a table
 arranged in a suitable format.
}
\description{
Print methods for objects of class PostHoc
}
\examples{
MM <- glm(Y ~ Treatment + 0,  data = DeIdentifiedExample)
GG <- posthoc(MM)
print(GG)
}
\author{
Rodrigo Labouriau
}
