% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods4PostHoc.R
\name{summary.PostHoc}
\alias{summary.PostHoc}
\title{Summary method for objects of class PostHoc}
\usage{
\method{summary}{PostHoc}(object, ...)
}
\arguments{
\item{object}{an object of class PostHoc to be printed.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a dataframe constructed in such a way that when printing this
 dataframe yields a table arranged in a suitable format. The summary,
 differently than the print method displays also the matrix of p-values
 of all the pairwise comparisons.
}
\description{
Summary method for objects of class PostHoc
}
\examples{
MM <- glm(Y ~ Treatment + 0,  data = DeIdentifiedExample)
GG <- posthoc(MM)
summary(GG)
}
\author{
Rodrigo Labouriau
}
