% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliarFunctions.R
\name{WaldPvalues}
\alias{WaldPvalues}
\title{Calculates a  Wald approximation for the p-values of pairwise comparisons}
\usage{
WaldPvalues(Effects, CovMatrix, padjust = "fdr")
}
\arguments{
\item{Effects}{a vector containing the effects}

\item{CovMatrix}{the covariance matrix of the effects}

\item{padjust}{method for correcting for multiple testing as in the
function p.adjust (default = "fdr", if NULL no adjustments is made)}
}
\value{
a vector of p-values.
}
\description{
Calculates a  Wald approximation for the p-values of pairwise comparisons
}
\author{
Rodrigo Labouriau
}
